More Windows crap...
Start menu shortcuts are now installed in Neverball instead of Games.

Added replay converter to the installer.  There's no special option for
that, it's installed unconditionally with the rest of the game.

Added launcher batch files for neverball, neverputt and mapc.  Along
with the replay converter and the wrapper convert-replays batch file,
these are all installed in a subdirectory called "bin".  It's a bit of a
mess, though;  the main binaries and dynamic libraries are still in the
top directory.  I (and maybe a handful of other people) still get a nice
clean directory to add to PATH and the rest can still carelessly click
the executable files placed in an obvious location and not see any
console windows flashing by.



git-svn-id: https://s.snth.net/svn/neverball/trunk@1084 78b8d119-cf0a-0410-b17c-f493084dd1d7