Restore original Escape behaviour if pause key is not Escape
The "original" Escape  key behaviour meant here is  "exit no questions
asked".  The  implementation is  actually "wrong",  based on  the fact
that the  game converts Escape  key presses to joystick  "exit" button
events.  Thus, the "pause key is  not Escape" test is performed in the
state buttn  handler when  in fact  we have no  idea where  the button
event  came from.   However, the  resulting behaviour  believed to  be
identical to that in 1.4.0 for both keyboard and joystick.


git-svn-id: https://s.snth.net/svn/neverball/trunk@2081 78b8d119-cf0a-0410-b17c-f493084dd1d7