Rework ball/util widget token policy
Instead of using large negative values for the on-screen keyboard and
score board widget tokens, this patch introduces a "GUI bit" (the low
bit of the most significant byte of a 4-byte integer) which is ORed
with small integer values to obtain the final token values.  This is
done mainly to allow some of these tokens to be ORed with each other.

git-svn-id: https://s.snth.net/svn/neverball/trunk@2460 78b8d119-cf0a-0410-b17c-f493084dd1d7