Rip out locale <-> UTF-8 filename conversions (and a direct iconv dep)
It seemed like a good idea at some point, but in practice it ended up
being a distracting layer of indirection and complexity trying to
support a rare (bordering on hypothetical) use case.  And it did not
work in Windows anyway.  So as an alternative, I propose this simple
rule: use a UTF-8 locale or stick to ASCII range in replay filenames.

git-svn-id: https://s.snth.net/svn/neverball/trunk@2814 78b8d119-cf0a-0410-b17c-f493084dd1d7