Add basic rotating objects (ODE only)
Rotating objects are represented in the map by func_rotating entities
(a la func_train).  First connected path_corner gives the center of
rotation, the "speed" attribute of this path_corner gives the time of
one full revolution.  If this path_corner is connected to another
path_corner, the directed line from one to the other gives the axis of
rotation (rotation follows the right hand rule).  If a second
path_corner is not connected, rotation axis is assumed to be the "up"
axis.

git-svn-id: https://s.snth.net/svn/neverball/trunk@3122 78b8d119-cf0a-0410-b17c-f493084dd1d7