Improve rotation behavior under constant centrifugal force
The ball had a tendency to slowly sink through walls when it was
"stably" leaning against them under centrifugal force. Rewrite
rotation calculations in a way that avoids this problem. There's still
at least one failure case though: when the path part of the rotating
body changes in the middle of the time step the result can be badly
off.

From: Uoti Urpala

git-svn-id: https://s.snth.net/svn/neverball/trunk@3321 78b8d119-cf0a-0410-b17c-f493084dd1d7