Add new billboard for marking platform paths
The aim is to have a path marker texture that has better visibility than poof-blue when used with light backgrounds (e.g., alien) while retaining a reasonable look in darker ones. The image has been softened slightly to make the edges appear less jagged when mipmapping is turned off, as is the default setting. (May be revised later.)

git-svn-id: https://s.snth.net/svn/neverball/trunk@3503 78b8d119-cf0a-0410-b17c-f493084dd1d7