Enforce requirement for pointer to hover clicked widget to activate
There should be no behavior changes from two commits ago. This commit
together with the last one simply ensures that there's always an
active widget, as in 1.5.4 and earlier. In turn, gui_start() is again
useful and joystick/keyboard users don't have to ponder the
existential question of "which widget is active when no widget is
active?"

git-svn-id: https://s.snth.net/svn/neverball/trunk@3959 78b8d119-cf0a-0410-b17c-f493084dd1d7