Fix mapc not to merge planes that are not perfectly coplanar
This fixes a problem that was discovered in map-mym2/littlecones.map a
really, really long time ago, where mapc seemingly had merged together
planes from two cones that were nearly-but-not-quite coplanar. This
caused a nasty collision artifact at the top of the cone in front of
the spawn point.

The downside of this patch is an about 10% increase in total SOL size.