Add point min and max size specification
Point sprite scaling is somewhat unpredictable, and varies from one
driver to the next. The most significant differences are the default
values giving the range of sizes at which a sprite can appear. So
we need to specify limits explicitly if we want consistency across all
platforms. Unfortunately, the order and time at which these values are
set appears to determine whether they take hold or not. This too varies
from one platform to the next.

This change takes control of the situation on Linux NVIDIA. Hopefully
it'll port. The maximum sprite size is set to one quarter of the height
of the screen (this is a value in pixels) and the default sizes of the
entity particles are given at around a tenth the height of the screen
(this is a value in pixels as viewed from a distance of 1 unit).

The distance attenuation factors give the change in the AREA of a point,
not the diameter (as do ALL the other values), so perspective-correct
point attenuation must use the quadratic coefficient, not the linear
one. This corrected interpretation is put to use for both coin particles
and material particles.