HMD fail gracefully on incapable hardware
This small change checks the framebuffer, vertex buffer, and shader object
state when asked to confirm HMD mode. This causes the game to fall back
correctly on non-HMD rendering if any of the above extensions to OpenGL
are missing.

We might want to add a menu element that highlights the circumstance where
CONFIG_HMD is true but hmd_stat() is false, as this would indicate that the
user tried to turn HMD support on, but the system failed. (Currently it runs
normally but the HMD option says "on", which may be confusing.)