Fix dark ring around green coin pad
The previous version of the green coin pad had black pixels in its transparent
areas. This causes a dark ring to appear around the coin pads when viewed from
a distance. This is due to mipmapping. When the texture is downsampled by the
driver, these normally-invisible black pixels end up contributing to visible
pixel averages. Thus blackness spreads as the renderer makes texel references
higher and higher up the mipmap hierarchy.

This is an art problem, not a code problem, and it's difficult to fix because
Photoshop INSISTS upon blacking (or whiting) out transparent pixels. The most
reliable means of exporting a PNG with correct color in its transparent areas
from Photoshop:

- Create a layer mask from the current transparency.
- Copy the contents of the layer mask.
- Create a new channel called "Alpha".
- Paste the mask into the alpha channel.
- Delete the layer mask.
- Save the image as 32-bit Targa, making sure to check "Alpha Channels".
- Use Imagemagick to convert the TGA to a PNG.

What a pain.