Emscripten: use navigator.storage API if available
Although we use IndexedDB as the backing store for the user folder,
the browser can still delete ("evict") this data in a low storage
space situation. I've lost replays and scores due to this. Luckily,
there's an API that we can use to request permanent storage for data,
so that it doesn't get evicted. It kind of works, but is not available
everywhere. Also, the only way to revoke it, once granted, is for the
user to delete the site data.