Added missing include in share/fetch_curl.h
Without this include, strdup (which is #defined as a custom function, dupe_string in share/common.h) would not have a visible definition and would be assumed to return an int type, which is 32-bit, and will therefore truncate the address on 64-bit platforms.

This fix makes the declaration visible by including share/common.h, thus preventing truncation and the carrying of an incorrect address pointer value.