Add "Back" GUI button to many screens
Browsers reserve Escape for exiting fullscreen, which makes it less
than ideal for backing out of menus, gameplay, etc.

This also adds a "Start" GUI button to the level start screen, because
a solitary GUI button can't be unfocused, making it impossible to start
the game by pressing A (because it activates the focused button).

Closes #118.